Move readonly helper files into their own dir
